+++ CHANGELOG.rst
@@ -2,6 +2,24 @@
 CHANGELOG
 =========
 
+0.7 (2021-04-10)
+------------------
+
+* refactor to use a :class:`BaseBot`
+* added a CLI
+* automatic generation of ssh key pairs and server setup
+* added a :class:`Deploybot`
+
+0.6 (2021-04-10)
+------------------
+
+* ...
+
+0.5 (2021-04-10)
+------------------
+
+* ...
+
 0.4.1 (2021-04-10)
 ------------------